src/consumer/requirements.txt
 confluent-kafka==1.8.2